pyproject.toml
[tool.poetry]
name = "github-analysis"
version = "0.1.0"
description = "Analyse research software repositories"
authors = ["James Graham <j.graham@software.ac.uk>"]
license = "BSD-3-Clause"

[tool.poetry.scripts]
gha = 'github_analysis.__main__:cli'

[tool.poetry.dependencies]
python = "^3.7"
pymongo = "^3.11.3"
requests = "^2.25.1"
python-decouple = "^3.4"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
prospector = "^1.3.1"
yapf = "^0.31.0"
pytest = "^6.2.3"

[tool.yapf]
COLUMN_LIMIT = 110

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
